requirements.txt
opencv-python~=4.5.3.56
numpy~=1.22.0
pillow
pypinyin~=0.43.0
yaml~=0.2.5
pyyaml~=6.0
prettytable~=2.2.1
tqdm~=4.62.3
requests~=2.26.0
unrar~=0.4
faker~=8.8.1
pandas~=1.3.3
mimesis~=4.1.3
vcam~=1.0
pyqt5~=5.15.4
wcwidth~=0.2.5
handright~=8.0.0
pyrect~=0.2.0
translatepy~=2.3